Remove ruby 2.1.2 requirement from Gemfile
I've been using Ruby 2.2.1, and the site builds fine. Clearly 2.1.2 is not a
strict requirement, and I'll bet that ruby >=2.1.2 works just fine. Stating it
as strict breaks bundler for other rubies, which is wrong. Just removing it for
now -- we can peg a set of supported rubies later, but presumably the evolving
gemset will define this amorphously.
Newer README (a TODO item) will state that this is a requirement we've removed,
in case it needs to be restored for some reason in the future. That will be
easier to find than git-blame on Gemfile, which is non-obvious.