Update deploy.sh to do s3 sync instead of cp
Using sync is cheaper network-wise and therefore faster. Importantly, doing a
delete sync after the upload removes any files that are unwanted from the
bucket.
It's safer to do two syncs, the second with --delete, than a single sync. It
guarantees that the deletions won't happen until after the upload is complete.