"Replace" python pygments with pygments.rb
This is just a ruby package that wraps the python pygments library, but it
removes the developer burden of handling a Gemfile and a virtualenv and keeping
everything happy. Results in much simplified instructions.
Also breaks down the deploy script into a scripts/ dir and a Makefile.
We now `make staging` and `make production`