Update Gemfile to handle eventmachine
eventmachine is a gem that is broken depending on which platform/version
pairing you use. Unfortunately, deep support for platform dispatch doesn't
exist in Gemfiles -- you can only do it for ruby versions if you want to
persist it to the Gemfile.lock. It is for this reason that I've discarded the
lock, which is wrong for someone's platform at any given time.
Instead, this logic has to live in the Gemfile as a dispatch on the platform
between 'darwin' and 'linux'
Comments in the Gemfile indicate what's going on.

As an add-on to this commit, also updates the setup instructions to include
installation of awscli, which was missing before. And moves 'make build' to be
a phony target.