Fix to asciidoc imagesdir -- remove trailing slash
Hotfixing this commit directly onto the production branch because this is
needed for the deployment in progress.

If asciidoc has a trailing slash in the 'imagesdir' value, it will generate
double-slash URLs like '..//images/abc.png'
On most webservers, that doesn't matter, but because the site is moving to S3,
and S3 is a blob-store in which '/' is a non-special identifier character,
'a//b' is a distinct valid identifier from 'a/b'. As a result, images from
asciidoc fail to load on S3 and pages are broken.
To fix, simply change 'imagesdir' to '..' instead of '../'