Updates to get Nanoc4 running minimally
Ultimately, a lot of the pain with the 3->4 upgrade comes from the
legacy identifier style not matching the new style of identifiers. Even
though they can be pattern matched in "legacy" mode, they don't have the
same API for manipulation, and we get a lot of pain. Rather than writing
a lot of shims to make legacy identifiers work in the new world, I've
chosen to move us onto the new Nanoc4 identifier style for cleanliness.
That means that a lot of these changes are a matter of identifier
compatibility, and that there are many files touched in small ways.

A few noteworthy changes that took place along the way:
- move /static/ into /content/static/ for clarity and ease
- make __TRANSFER_VERSION__ transformation into a nanoc filter
- add some nice helpers for use in the Rules
- remove preprocess from Rules
- allow dots in identifiers (made some things work better, like patterns
  matching on extension)

Rewrote the Gemfile to be much simpler, listing a very small set of
dependencies.