Fold-down specific layouts, fix title magick
Title detection from adoc files needs to work because it's used in a
bunch of layouts for a visible page title.
Collapses down from many section-specific layouts (CLI, Auth, Transfer,
FAQ, etc) to a more generic layout for pages with a sidebar, and a
helper utility for getting the variable information out of the item
being laid out.