Fix all menu rendering
That's right, all of it.
Also removes some old garbage from the lib/ dir.
Required a full rewrite of some of the core menu rendering logic, during
which I folded together the sidebar and regular menu structures. Now we
just "render_menu", potentially in sidebar mode. That removes a lot of
redundant logic. The parity between the two methods was a headache
because they would often encode the same logic two different ways --
having two modes for one method turns out to be a lot simpler to read
and understand.