Fix 'active' class in sidebar menu
In the sidebar, we want to highlight the current item when it's selected
in the menu. This fixes the logic that handles that to check if we're
looking at the current item explicitly, which is a lot simpler than the
murky path based matching which we were doing.
Also fixes `GlobusNavigation::safe_item_title` to look at the
:short_title if it's set.