Rejigger caret class for sidebar
Caret class, Y U SO CARET?
Some actual detail:
We were correctly handling this in the case of nested sidebar menus (of
which we have none), detecting that we're on a page which should be
expanded as part of the current menu and changing the caret class
appropriately. However, at the top level of such menus -- i.e. at the
level of /cli/reference/ above /cli/reference/acl-add/ -- we were not
changing the type of caret appropriately. As a result, the caret was
pointing the wrong way around, and toggling backwards as well. After the
fix, we percolate the caret class change up one level higher than
before, correctly capturing the top level of nested sidebar menus.